nifty_intraday_strategy/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
pyyaml>=6.0
scipy>=1.9.0

